Refactor: move taktikal import to its own file and rename some things
